Add %{severity} and %s formatters
I've become used to seeing the severity level as a single letter at the beginning of the
log line. Make it possible here with %s, and make it the default.  %{severity} can be
used for the long severity level name: INFO, WARNING, etc.

Changed WARN to WARNING as that always seemed too arbitrary to me and was due to wanting
to keep all the severity names short-ish. Made all the names uppercase because the
inconsistency was driving me crazy.