Make `$stdout-log-target` work when `*standard-output*` is rebound
This is useful because, for example, the http libraries do a lot of log output to the
console during testing. Testworks has a general mechanims for capturing output from tests
by rebinding `*standard-output*` and `*standard-error*`, and that only works if the log
target didn't capture the value of one of those variables when it was created.